requirements.txt
langchain
langchain-community
langchain-openai
langchain-groq
python-dotenv
langchain-google-genai
streamlit
langchain-core
openai